+++ SpatialGDK/Source/SpatialGDKEditor/Private/SpatialGDKEditorCloudLauncher.cpp
@@ -18,7 +18,7 @@ bool SpatialGDKCloudLaunch()
 	const USpatialGDKEditorSettings* SpatialGDKSettings = GetDefault<USpatialGDKEditorSettings>();
 
 	FString LauncherCreateArguments = FString::Printf(
-		TEXT("create %s %s %s %s \"%s\" \"%s\" %s \"%s\""),
+		TEXT("create %s %s %s %s \"%s\" \"%s\" %s \"%s\" \"%s\""),
 		*FSpatialGDKServicesModule::GetProjectName(),
 		*SpatialGDKSettings->GetAssemblyName(),
 		*SpatialGDKSettings->GetSpatialOSRuntimeVersionForCloud(),
@@ -26,7 +26,8 @@ bool SpatialGDKCloudLaunch()
 		*SpatialGDKSettings->GetPrimaryLaunchConfigPath(),
 		*SpatialGDKSettings->GetSnapshotPath(),
 		*SpatialGDKSettings->GetPrimaryRegionCode().ToString(),
-		*SpatialGDKSettings->GetMainDeploymentCluster()
+		*SpatialGDKSettings->GetMainDeploymentCluster(),
+		*SpatialGDKSettings->GetDeploymentTags()
 	);
 
 	if (SpatialGDKSettings->IsSimulatedPlayersEnabled())