+++ CHANGELOG.md
@@ -85,6 +85,7 @@ Usage: `DeploymentLauncher createsim <project-name> <assembly-name> <target-depl
 - A launch configuration editor has been added under the Deploy toolbar button.
 - The cloud deployment window can now generate a launch configuration from the current map or use the launch configuration editor.
 - Worker load can be specified by game logic via `SpatialMetrics::SetWorkerLoadDelegate`
+- You can now specify deployment tags in the `Cloud Deployment` window.
 
 ## Bug fixes:
 - Fixed a bug that caused queued RPCs to spam logs when an entity is deleted.